+++ test/hemera/client-extension-errors.spec.js
@@ -1,0 +1,217 @@
+'use strict'
+
+describe('Client Extension error', function() {
+  var PORT = 6242
+  var authUrl = 'nats://localhost:' + PORT
+  var server
+
+  // Start up our own nats-server
+  before(function(done) {
+    server = HemeraTestsuite.start_server(PORT, done)
+  })
+
+  // Shutdown our server after we are done
+  after(function() {
+    server.kill()
+  })
+
+  it('Should be able to pass a super error to onClientPostRequest', function(
+    done
+  ) {
+    const nats = require('nats').connect(authUrl)
+
+    const hemera = new Hemera(nats)
+
+    let plugin = function(options) {
+      let hemera = this
+
+      hemera.ext('onClientPostRequest', function(ctx, next) {
+        next(new UnauthorizedError('test'))
+      })
+
+      hemera.add(
+        {
+          cmd: 'add',
+          topic: 'math'
+        },
+        (resp, cb) => {
+          cb(null, resp.a + resp.b)
+        }
+      )
+    }
+
+    hemera.use({
+      plugin: plugin,
+      attributes: {
+        name: 'myPlugin'
+      },
+      options: {}
+    })
+
+    hemera.ready(() => {
+      hemera.act(
+        {
+          topic: 'math',
+          cmd: 'add',
+          a: 1,
+          b: 2
+        },
+        (err, resp) => {
+          expect(err).to.be.exists()
+          expect(err.name).to.be.equals('Unauthorized')
+          expect(err.message).to.be.equals('test')
+          hemera.close(done)
+        }
+      )
+    })
+  })
+
+  it('Should be able to pass an error to onClientPostRequest', function(done) {
+    const nats = require('nats').connect(authUrl)
+
+    const hemera = new Hemera(nats)
+
+    let plugin = function(options) {
+      let hemera = this
+
+      hemera.ext('onClientPostRequest', function(ctx, next) {
+        next(new Error('test'))
+      })
+
+      hemera.add(
+        {
+          cmd: 'add',
+          topic: 'math'
+        },
+        (resp, cb) => {
+          cb(null, resp.a + resp.b)
+        }
+      )
+    }
+
+    hemera.use({
+      plugin: plugin,
+      attributes: {
+        name: 'myPlugin'
+      },
+      options: {}
+    })
+
+    hemera.ready(() => {
+      hemera.act(
+        {
+          topic: 'math',
+          cmd: 'add',
+          a: 1,
+          b: 2
+        },
+        (err, resp) => {
+          expect(err).to.be.exists()
+          expect(err.name).to.be.equals('Error')
+          expect(err.message).to.be.equals('test')
+          hemera.close(done)
+        }
+      )
+    })
+  })
+
+  it('Should be able to pass a custom super error to onClientPreRequest', function(
+    done
+  ) {
+    const nats = require('nats').connect(authUrl)
+
+    const hemera = new Hemera(nats)
+
+    let plugin = function(options) {
+      let hemera = this
+
+      hemera.ext('onClientPreRequest', function(ctx, next) {
+        next(new UnauthorizedError('test'))
+      })
+
+      hemera.add(
+        {
+          cmd: 'add',
+          topic: 'math'
+        },
+        (resp, cb) => {
+          cb(null, resp.a + resp.b)
+        }
+      )
+    }
+
+    hemera.use({
+      plugin: plugin,
+      attributes: {
+        name: 'myPlugin'
+      },
+      options: {}
+    })
+
+    hemera.ready(() => {
+      hemera.act(
+        {
+          topic: 'math',
+          cmd: 'add',
+          a: 1,
+          b: 2
+        },
+        (err, resp) => {
+          expect(err).to.be.exists()
+          expect(err.name).to.be.equals('Unauthorized')
+          expect(err.message).to.be.equals('test')
+          hemera.close(done)
+        }
+      )
+    })
+  })
+
+  it('Should be able to pass an error to onClientPreRequest', function(done) {
+    const nats = require('nats').connect(authUrl)
+
+    const hemera = new Hemera(nats)
+
+    let plugin = function(options) {
+      let hemera = this
+
+      hemera.ext('onClientPreRequest', function(ctx, next) {
+        next(new Error('test'))
+      })
+
+      hemera.add(
+        {
+          cmd: 'add',
+          topic: 'math'
+        },
+        (resp, cb) => {
+          cb(null, resp.a + resp.b)
+        }
+      )
+    }
+
+    hemera.use({
+      plugin: plugin,
+      attributes: {
+        name: 'myPlugin'
+      },
+      options: {}
+    })
+
+    hemera.ready(() => {
+      hemera.act(
+        {
+          topic: 'math',
+          cmd: 'add',
+          a: 1,
+          b: 2
+        },
+        (err, resp) => {
+          expect(err).to.be.exists()
+          expect(err.name).to.be.equals('Error')
+          expect(err.message).to.be.equals('test')
+          hemera.close(done)
+        }
+      )
+    })
+  })
+})
